feat(onboarding): add context provider
