+++ src/features/Onboarding/pages/EnterpriseOnboardWizard.tsx
@@ -1,18 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { ChevronRight, Building2, User, MapPin, Phone, Check, Loader2, FileText } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { cn } from '@/lib/utils';
-import supabase from '@/utils/supabase/client';
-import { Tables, TablesInsert } from '@/utils/supabase/types';
-
-// Type aliases for better readability
-type BrazilianState = Tables<"brazilian_states">;
-type BusinessSegment = Tables<"business_segments">;
-type OrganizerType = Tables<"organizer_types">;
-type SubscriptionPlan = Tables<"subscription_plans">;
-type TenantStatus = Tables<"tenant_statuses">;
-type TenantInsert = TablesInsert<"tenants">;
+import { useOnboarding, OnboardingProvider } from '../context/OnboardingContext';
 
 interface StepIndicatorProps {
   step: number;
@@ -55,248 +46,32 @@ const StepIndicator: React.FC<StepIndicatorProps> = ({ step, currentStep, title,
   );
 };
 
-interface FormData {
-  // User data
-  email: string;
-  password: string;
-  
-  // Company basic info
-  razaoSocial: string;
-  nomeFantasia: string;
-  cnpj: string;
-  inscricaoEstadual: string;
-  cnaePrincipal: string;
-  segmentoId: string;
-  organizerTypeId: string;
-  
-  // Contact info
-  contactEmail: string;
-  phone: string;
-  whatsapp: string;
-  website: string;
-  
-  // Location
-  logradouro: string;
-  numero: string;
-  complemento: string;
-  bairro: string;
-  cep: string;
-  estadoId: string;
-  cidade: string;
-}
 
 const EnterpriseOnboardWizard: React.FC = () => {
-  const [currentStep, setCurrentStep] = useState(1);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [states, setStates] = useState<BrazilianState[]>([]);
-  const [segments, setSegments] = useState<BusinessSegment[]>([]);
-  const [organizerTypes, setOrganizerTypes] = useState<OrganizerType[]>([]);
-  const [plans, setPlans] = useState<SubscriptionPlan[]>([]);
-  const [defaultPlanId, setDefaultPlanId] = useState<string>('00ace61e-0853-4254-be82-b888e91ce3c4');
-  const [defaultStatusId, setDefaultStatusId] = useState<string>('cbfa8c27-aa51-4a64-9e71-3d068728d9f5');
-  
-  const [formData, setFormData] = useState<FormData>({
-    email: '',
-    password: '',
-    razaoSocial: '',
-    nomeFantasia: '',
-    cnpj: '',
-    inscricaoEstadual: '',
-    cnaePrincipal: '',
-    segmentoId: '',
-    organizerTypeId: '',
-    contactEmail: '',
-    phone: '',
-    whatsapp: '',
-    website: '',
-    logradouro: '',
-    numero: '',
-    complemento: '',
-    bairro: '',
-    cep: '',
-    estadoId: '',
-    cidade: ''
-  });
-
-  // Load all necessary data on mount
-  useEffect(() => {
-    const loadData = async () => {
-      try {
-        // Load states
-        const { data: statesData } = await supabase
-          .from('brazilian_states')
-          .select('*')
-          .eq('is_active', true)
-          .order('name');
-        if (statesData) setStates(statesData);
-
-        // Load business segments
-        const { data: segmentsData } = await supabase
-          .from('business_segments')
-          .select('*')
-          .eq('is_active', true)
-          .order('sort_order');
-        if (segmentsData) setSegments(segmentsData);
-
-        // Load organizer types
-        const { data: organizerData } = await supabase
-          .from('organizer_types')
-          .select('*')
-          .eq('is_active', true)
-          .order('sort_order');
-        if (organizerData) setOrganizerTypes(organizerData);
-
-        // Load subscription plans
-        const { data: plansData } = await supabase
-          .from('subscription_plans')
-          .select('*')
-          .eq('is_active', true)
-          .order('sort_order');
-        if (plansData) {
-          setPlans(plansData);
-          // Set default plan (usually the free/trial plan)
-          const trialPlan = plansData.find(p => p.code === 'trial' || p.code === 'free');
-          if (trialPlan) setDefaultPlanId(trialPlan.id);
-        }
-
-        // Load default status (active)
-        const { data: statusData } = await supabase
-          .from('tenant_statuses')
-          .select('id')
-          .eq('code', 'active')
-          .single();
-        if (statusData) setDefaultStatusId(statusData.id);
-      } catch (error) {
-        console.error('Error loading data:', error);
-      }
-    };
-    loadData();
-  }, []);
-
-  const updateFormData = (field: keyof FormData, value: string) => {
-    setFormData(prev => ({ ...prev, [field]: value }));
-  };
-
-  const generateSlug = (razaoSocial: string): string => {
-    return razaoSocial
-      .toLowerCase()
-      .normalize('NFD')
-      .replace(/[\u0300-\u036f]/g, '') // Remove accents
-      .replace(/[^a-z0-9]+/g, '-') // Replace non-alphanumeric with hyphens
-      .replace(/^-+|-+$/g, '') // Remove leading/trailing hyphens
-      .substring(0, 50); // Max 50 chars
-  };
+  const {
+    currentStep,
+    isSubmitting,
+    states,
+    segments,
+    organizerTypes,
+    plans,
+    formData,
+    nextStep,
+    prevStep,
+    updateFormData,
+    submit,
+  } = useOnboarding();
 
   const handleNext = () => {
     if (currentStep < 5) {
-      setCurrentStep(currentStep + 1);
+      nextStep();
     } else {
-      handleSubmit();
+      submit();
     }
   };
 
   const handlePrev = () => {
-    if (currentStep > 1) {
-      setCurrentStep(currentStep - 1);
-    }
-  };
-
-  const handleSubmit = async () => {
-    setIsSubmitting(true);
-    try {
-      // Create user account
-      const { data: authData, error: authError } = await supabase.auth.signUp({
-        email: formData.email,
-        password: formData.password,
-      });
-
-    //   if (authError) throw authError;
-
-      // Generate slug from razao social
-      const slug = generateSlug(formData.razaoSocial);
-
-      // Calculate trial end date (30 days from now)
-      const trialEndsAt = new Date();
-      trialEndsAt.setDate(trialEndsAt.getDate() + 30);
-
-      // Create tenant record with proper types
-      const tenantData: TenantInsert = {
-        slug: slug,
-        cnpj: formData.cnpj,
-        razao_social: formData.razaoSocial,
-        nome_fantasia: formData.nomeFantasia || null,
-        inscricao_estadual: formData.inscricaoEstadual || null,
-        cnae_principal: formData.cnaePrincipal || null,
-        organizer_type_id: formData.organizerTypeId,
-        primary_segment_id: formData.segmentoId,
-        email_domain: formData.contactEmail.split('@')[1] || null,
-        status_id: defaultStatusId,
-        plan_id: defaultPlanId,
-        trial_ends_at: trialEndsAt.toISOString(),
-        contact_email: formData.contactEmail,
-        contact_phone: formData.phone || null,
-        whatsapp_number: formData.whatsapp || null,
-        website_url: formData.website || null,
-        endereco_logradouro: formData.logradouro,
-        endereco_numero: formData.numero || null,
-        endereco_complemento: formData.complemento || null,
-        endereco_bairro: formData.bairro,
-        endereco_cidade: formData.cidade,
-        state_id: formData.estadoId,
-        cep: formData.cep,
-        created_by: authData.user.id,
-        onboarding_current_step: 'dados_empresa',
-        lgpd_acceptance_date: new Date().toISOString(),
-        // Default values as per DDL
-        primary_color: '#4D2BFB',
-        secondary_color: '#03F9FF',
-        font_family: 'Neue Haas Unica',
-        timezone: 'America/Sao_Paulo',
-        locale: 'pt-BR',
-        payment_method: 'pix',
-        domain_validated: false,
-        current_events_count: 0,
-        current_admins_count: 0,
-        max_events_allowed: 1,
-        max_admins_allowed: 1,
-        max_visitors_allowed: 100,
-        max_exhibitors_allowed: 10,
-        onboarding_completed: false,
-        setup_wizard_completed: false,
-        first_event_created: false,
-        features_enabled: {},
-        integrations_config: {},
-        data_retention_months: 24,
-        optante_simples_nacional: false,
-        total_events_created: 0,
-        total_revenue_brl: 0
-      };
-
-      const { data: tenantResult, error: tenantError } = await supabase
-        .from('tenants')
-        .insert([tenantData])
-        .select()
-        .single();
-
-      if (tenantError) throw tenantError;
-
-      // Create user-tenant association if you have a tenant_users table
-      if (tenantResult) {
-        // This assumes you have a tenant_users table
-        await supabase
-          .from('tenants')
-          .insert([tenantData]);
-      }
-
-      alert('Cadastro realizado com sucesso! Verifique seu email para confirmar a conta.');
-      // Redirect to login or dashboard
-      window.location.href = '/login';
-    } catch (error) {
-      console.error('Error submitting form:', error);
-      alert('Erro ao realizar cadastro. Verifique os dados e tente novamente.');
-    } finally {
-      setIsSubmitting(false);
-    }
+    prevStep();
   };
 
   const formatCNPJ = (value: string) => {
@@ -764,4 +539,10 @@ const EnterpriseOnboardWizard: React.FC = () => {
   );
 };
 
-export default EnterpriseOnboardWizard;
+const EnterpriseOnboardWizardPage: React.FC = () => (
+  <OnboardingProvider>
+    <EnterpriseOnboardWizard />
+  </OnboardingProvider>
+);
+
+export default EnterpriseOnboardWizardPage;
